% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_mt.R
\name{create_mt}
\alias{create_mt}
\title{create a machine tag}
\usage{
create_mt(
  datasetkey,
  namespace,
  name,
  value,
  embedValueList = FALSE,
  api = "gbif"
)
}
\arguments{
\item{datasetkey}{uuid}

\item{namespace}{the name of the machinetag in general}

\item{name}{the name of the machineTag often similar to namespace}

\item{value}{the value of the machine tag}

\item{embedValueList}{use this to embed json into a string}

\item{api}{"gbif" or "uat"}
}
\value{
Status message.
}
\description{
create a machine tag
}
\examples{
\dontrun{
create_mt(
datasetkey="38b4c89f-584c-41bb-bd8f-cd1def33e92f",
namespace="testMachineTag.jwaller.gbif.org",
name="testMachineTag",
value=1,
api="uat")
}

}
