% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_mt.R
\name{delete_mt}
\alias{delete_mt}
\title{Delete a machine tag}
\usage{
delete_mt(datasetkey, key, api = "gbif")
}
\arguments{
\item{datasetkey}{uuid}

\item{key}{key of machine tag}

\item{api}{gbif or uat}
}
\value{
a message
}
\description{
Delete a machine tag
}
\examples{
\dontrun{
delete_mt("38b4c89f-584c-41bb-bd8f-cd1def33e92f",6229107,api="uat")
}
}
