% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mt.R
\name{get_mt}
\alias{get_mt}
\title{Get a machine tag}
\usage{
get_mt(
  machineTagNamespace = NULL,
  machineTagValue = NULL,
  machineTagName = NULL,
  limit = 500,
  api = "gbif",
  type = "dataset"
)
}
\arguments{
\item{machineTagNamespace}{namespace}

\item{machineTagValue}{value}

\item{machineTagName}{name}

\item{api}{gbif or uat}

\item{type}{"organization", "dataset", "installation"}
}
\value{
`tibble` of machine tag data.
}
\description{
Get a machine tag
}
\examples{
\dontrun{

get_mt(machineTagNamespace="testMachineTag.jwaller.gbif.org",api="uat")

}
}
